✨ Add delete file mutation as its own component
