+++ src/containers/FileActionsContainer.jsx
@@ -1,47 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Mutation} from 'react-apollo';
-import {GET_STUDY_BY_ID} from '../state/queries';
-import {DELETE_FILE, FILE_DOWNLOAD_URL} from '../state/mutations';
+import {FILE_DOWNLOAD_URL} from '../state/mutations';
 import FileActionButtons from '../components/FileActionButtons/FileActionButtons';
-
+import DeleteFileMutation from './DeleteFileMutation';
 const FileActionsContainer = ({node, studyId, className}) => {
   if (node) {
     return (
       <Mutation mutation={FILE_DOWNLOAD_URL} key={node.kfId}>
         {downloadFile => (
-          <Mutation
-            mutation={DELETE_FILE}
-            update={(cache, {data: {deleteFile}}) => {
-              // Re-writes the study in the cache with the deleted file
-              // removed.
-              const {studyByKfId} = cache.readQuery({
-                query: GET_STUDY_BY_ID,
-                variables: {kfId: studyId},
-              });
-              const deleteIndex = studyByKfId.files.edges.findIndex(
-                edge => edge.node.kfId === deleteFile.kfId,
-              );
-              if (deleteIndex < 0) {
-                return studyByKfId;
-              }
-              studyByKfId.files.edges.splice(deleteIndex, 1);
-              const data = {
-                studyByKfId: {
-                  ...studyByKfId,
-                  files: {
-                    __typename: 'FileNodeConnection',
-                    edges: studyByKfId.files.edges,
-                  },
-                },
-              };
-              cache.writeQuery({
-                query: GET_STUDY_BY_ID,
-                data,
-              });
-            }}
-            key={node.kfId}
-          >
+          <DeleteFileMutation studyId={studyId}>
             {(deleteFile, {loading, error}) => (
               <FileActionButtons
                 {...{
@@ -55,7 +23,7 @@ const FileActionsContainer = ({node, studyId, className}) => {
                 }}
               />
             )}
-          </Mutation>
+          </DeleteFileMutation>
         )}
       </Mutation>
     );
